requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
python-multipart==0.0.12
pydantic==2.9.2
pydantic-settings==2.6.0
celery==5.4.0
redis==5.0.8
Jinja2==3.1.4
python-jose[cryptography]==3.3.0
httpx==0.27.2
prometheus-fastapi-instrumentator==6.1.0
email-validator==2.2.0
pyyaml==6.0.2
psutil==6.0.0
pyannote.audio==3.3.2
soundfile==0.12.1
numpy==1.26.4
faster-whisper==1.0.3
python-dotenv==1.0.1
reportlab==4.2.5
openai-whisper==20231117
